# Conflicts:
#	src/main/java/com/umc/cardify/controller/FolderController.java
#	src/main/java/com/umc/cardify/dto/folder/FolderResponse.java
#	src/main/java/com/umc/cardify/service/FolderService.java
#	src/main/java/com/umc/cardify/service/NoteComponentService.java

+++ src/main/java/com/umc/cardify/service/NoteComponentService.java
@@ -23,13 +23,26 @@ import com.umc.cardify.config.exception.BadRequestException;
 import com.umc.cardify.config.exception.DatabaseException;
 import com.umc.cardify.config.exception.ErrorResponseStatus;
 import com.umc.cardify.converter.NoteConverter;
+import com.umc.cardify.domain.*;
 import com.umc.cardify.domain.ProseMirror.Node;
 import com.umc.cardify.domain.enums.MarkStatus;
+import com.umc.cardify.domain.enums.SubscriptionStatus;
 import com.umc.cardify.dto.note.NoteRequest;
 import com.umc.cardify.dto.note.NoteResponse;
-
+import com.umc.cardify.repository.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.time.LocalDateTime;
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Slf4j
 @Service
@@ -58,9 +71,9 @@ public class NoteComponentService {
 			Note newNote = NoteConverter.toAddNote(folder);
 			noteRepository.save(newNote);
 			try {
-				ContentsNote contentsNote = ContentsNote.builder().noteId(newNote.getNoteId()).build();
-				contentsNoteRepository.insert(contentsNote);
-				newNote.setContentsId(contentsNote.get_id());
+				ContentsNote contentsNote = ContentsNote.builder().note(newNote).build();
+				contentsNoteRepository.save(contentsNote);
+				newNote.setContentsNote(contentsNote);
 			}catch (Exception e){
 				System.out.println(e);
 			}
@@ -94,7 +107,7 @@ public class NoteComponentService {
 			throw new BadRequestException(ErrorResponseStatus.INVALID_USERID);
 		else {
 			noteRepository.delete(note_del);
-			contentsNoteRepository.delete(contentsNoteRepository.findByNoteId(note_del.getNoteId()).get());
+			contentsNoteRepository.delete(contentsNoteRepository.findByNote(note_del).get());
 			return true;
 		}
 	}
@@ -156,6 +169,13 @@ public class NoteComponentService {
 
 	@Transactional
 	public Boolean writeNote(NoteRequest.WriteNoteDto request, Long userId, List<MultipartFile> images) {
+        // 작성 모드 설정
+        String mode = request.getMode();
+        if(mode == null || mode.isEmpty())
+            mode = "standard";
+        if(!mode.equals("standard") && !mode.equals("light"))
+            throw new BadRequestException(ErrorResponseStatus.REQUEST_ERROR);
+
 		Note note = noteModuleService.getNoteById(request.getNoteId());
 
 		if (!userId.equals(note.getFolder().getUser().getUserId())) {
@@ -166,25 +186,35 @@ public class NoteComponentService {
 			log.warn("IsEdit is : {}", note.getIsEdit());
 			throw new BadRequestException(ErrorResponseStatus.DB_UPDATE_ERROR);
 		}
-		if (cardModuleService.existsByNote(note)) {
+		if (cardModuleService.existsByNote(note) && mode.equals("standard")) {
 			cardModuleService.deleteAllCardsByNoteId(note.getNoteId());
 			cardModuleService.deleteAllImageCardsByNoteId(note.getNoteId());
-
 		}
+        note.setName(request.getName());
+        Node node = request.getContents();
 
-		StringBuilder totalText = new StringBuilder();
-		note.setName(request.getName());
+        if(mode.equals("standard")) {
+            StringBuilder totalText = new StringBuilder();
 
-		Queue<MultipartFile> imageQueue = new LinkedList<>(images != null ? images : Collections.emptyList());
-		Node node = request.getContents();
-		searchCard(node, totalText, note, imageQueue);
-		note.setTotalText(totalText.toString());
+            Queue<MultipartFile> imageQueue = new LinkedList<>(images != null ? images : Collections.emptyList());
+            searchCard(node, totalText, note, imageQueue);
+            note.setTotalText(totalText.toString());
+        }
 
-		ContentsNote contentsNote = contentsNoteRepository.findByNoteId(note.getNoteId()).get();
-		contentsNote.setContents(node);
+		ContentsNote contentsNote = contentsNoteRepository.findByNote(note)
+                .orElseThrow(() -> new BadRequestException(ErrorResponseStatus.NOT_FOUND_ERROR));
+
+        String content;
+        try {
+            content = objectMapper.writeValueAsString(node);
+        } catch (JsonProcessingException e){
+            throw new BadRequestException(ErrorResponseStatus.REQUEST_ERROR);
+        }
+		contentsNote.setContents(content);
 		contentsNoteRepository.save(contentsNote);
 
 		noteModuleService.saveNote(note);
+
 		return true;
 	}
 
@@ -354,7 +384,7 @@ public class NoteComponentService {
 	}
 
 	public NoteResponse.getNoteDTO getNote(Long noteId) {
-		Note note = noteRepository.findById(noteId)
+        Note note = noteRepository.findById(noteId)
 			.orElseThrow(() -> new BadRequestException(ErrorResponseStatus.NOT_FOUND_ERROR));
 		//노트 조회 시간 갱신
 		note.setViewAt(LocalDateTime.now());
@@ -370,7 +400,7 @@ public class NoteComponentService {
 				.build();
 		}).toList();
 
-		return noteConverter.getNoteDTO(note, cardDTO);
+        return noteConverter.getNoteDTO(note, cardDTO);
 	}
 
 	public List<NoteResponse.NoteInfoDTO> getRecentNotes(Long userId, int page, Integer size) {
